rework gar-auth (-> oci-auth): support GHCR
It showed to be helpful for forked repositories to avoid pushing to
Gardener's Google-Artifact-Registry (such as to avoid having to grant
permissions, and also to avoid interference w/ contents published from
pipelines in upstream project). Since GitHub offers a built-in
OCI-Registry (GHCR), it is deemed a reasonable choice to by default
publish to GHCR-repositories for respective forked repository.

Hence, extend gar-auth action to also support authentication against
GHCR. Determine correct target to authenticate against based on
oci-image-reference.

As a downside, the action now requires github-access-token to also be
passed-in.